requirements.txt
flask
flask-login
flask-sqlalchemy
bcrypt
PyJWT
flask-jwt-extended
functools